requirements.txt
matplotlib
tkintermapview
geopy
openrouteservice
python-dotenv 